--- packages/common/src/utils/adminDB.ts
@@ -1,5 +1,0 @@
-import * as admin from "firebase-admin";
-
-const adminDB = admin.firestore();
-
-export default adminDB;
