+++ packages/functions/tsconfig.json
@@ -18,6 +18,7 @@
   "include": ["src/**/*"],
   "exclude": [
     "node_modules",
+    "**/__mocks__/**",
     "**/__tests__/*.ts",
     "**/__tests__/*.tsx",
     "**/*.spec.ts",
